requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
pytest-django==4.8.0
sqlparse==0.4.4
